+++ test/extensions/fraction_string_test.dart
@@ -8,6 +8,7 @@ void main() {
       expect('-4/5'.toFraction(), equals(Fraction(-4, 5)));
       expect('5'.toFraction(), equals(Fraction(5)));
       expect('-5'.toFraction(), equals(Fraction(-5)));
+      expect('⅘'.toFraction(), equals(Fraction(4, 5)));
     });
 
     test('Making sure that invalid strings conversions throw and exception',
@@ -22,8 +23,9 @@ void main() {
     test(
         'Making sure that the boolean check is safe to be used before converting',
         () {
-      expect('3/5'.isFraction(), isTrue);
-      expect(''.isFraction(), isFalse);
+      expect('3/5'.isFraction, isTrue);
+      expect('⅜'.isFraction, isTrue);
+      expect(''.isFraction, isFalse);
     });
   });
 }