+++ test/extensions/mixed_num_test.dart
@@ -4,20 +4,27 @@ import 'package:fraction/fraction.dart';
 void main() {
   group("Testing the extension method on 'num'", () {
     test(
-        'Making sure that integers are properly converted into mixed fractions',
-        () {
-      const integerVal = 16;
-      final mixedVal = MixedFraction(whole: 0, numerator: 16, denominator: 1);
+        'Making sure that integers are properly converted into '
+        'mixed fractions', () {
+      final mixedVal = MixedFraction(
+        whole: 0,
+        numerator: 16,
+        denominator: 1,
+      );
 
-      expect(integerVal.toMixedFraction(), equals(mixedVal));
+      expect(16.toMixedFraction(), equals(mixedVal));
     });
 
-    test('Making sure that doubles are properly converted into mixed fractions',
-        () {
-      const doubleVal = 6.74;
-      final mixedVal = MixedFraction(whole: 6, numerator: 37, denominator: 50);
+    test(
+        'Making sure that doubles are properly converted into '
+        'mixed fractions', () {
+      final mixedVal = MixedFraction(
+        whole: 6,
+        numerator: 37,
+        denominator: 50,
+      );
 
-      expect(doubleVal.toMixedFraction(), equals(mixedVal));
+      expect(6.74.toMixedFraction(), equals(mixedVal));
     });
   });
 }